added API search and created MovieList component - Tile
