+++ src/Routes/Movies/index.js
@@ -1,28 +1,39 @@
 import {
-    StyledTitle, StyledMovies, StyledMovieLink, StyledImg,
-    StyledName, StyledYear, StyledGenres, StyledRate,
-    StyledStar, StyledAvarage, StyledVotes, StyledGenre,
-    StyledDetails
+    StyledTitle, StyledMovies
 } from "./styled";
-import star from "../../Images/star.svg";
-import { useSelector } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { selectMovieList, selectMovieListStatus } from "../../features/moveList/movieListSlice";
 import { selectGenreList, selectGenreListStatus } from "../../features/genreList/genreListSlice";
 import { Error } from "../../core/status/Error";
 import { Loading } from "../../core/status/Loading";
-import { Outlet } from "react-router-dom";
+import { Outlet, useLocation } from "react-router-dom";
+import { searchQueryParamName } from "../../features/useQueryParameter";
+import { useEffect } from "react";
+import { axiosSearchParamsMovie, selectSearchParamsMovieList } from "../../features/searchParams/searchParamsSlice";
+import Tile from "./Tile";
 
 const Movies = () => {
+    const dispatch = useDispatch();
+    const location = useLocation();
+
     const fetchData = useSelector(selectMovieList);
     const fetchGenre = useSelector(selectGenreList);
+    const fetchMoviesSearch = useSelector(selectSearchParamsMovieList);
+    const loadingMoviesSearch = useSelector(selectMovieListStatus);
     const loadingGeners = useSelector(selectGenreListStatus);
     const loadingMovies = (useSelector(selectMovieListStatus));
-    const URL = "https://www.themoviedb.org/t/p/w440_and_h660_face/";
+
+    const searchParams = (new URLSearchParams(location.search)).get(searchQueryParamName);
+    useEffect(() => {
+        dispatch(axiosSearchParamsMovie(searchParams))
+
+    }, [searchParams])
 
     if (loadingMovies === "success" && loadingGeners === "success") {
+        const searchList = fetchMoviesSearch !== null ? fetchMoviesSearch.results : null;
+
         const moviesList = fetchData.results;
         const genreList = fetchGenre.genres;
-        const genres = [];
 
         return (
             <>
@@ -30,36 +41,7 @@ const Movies = () => {
                     <StyledTitle>
                         Popular movies
                     </StyledTitle>
-                    {moviesList.map(movie => {
-                        return (
-                            <StyledMovieLink key={movie.id} to={`${movie.id}`}>
-                                <StyledImg src={`${URL}${movie.poster_path}`} />
-                                <StyledDetails>
-                                    <StyledName>
-                                        {movie.original_title}
-                                    </StyledName>
-                                    <StyledYear>
-                                        {(movie.release_date).slice(0, 4)}
-                                    </StyledYear>
-                                    <StyledGenres>
-                                        {(movie.genre_ids).forEach((element, index) => {
-                                            genres[index] = (genreList.filter(({ id }) => id === element)).map(({ name }) => name)
-                                        })}
-                                        {genres.map((list, index) => <StyledGenre key={index}>{list}</StyledGenre>)}
-                                    </StyledGenres>
-                                    <StyledRate>
-                                        <StyledStar src={star} />
-                                        <StyledAvarage>
-                                            {movie.vote_average}
-                                        </StyledAvarage>
-                                        <StyledVotes>
-                                            {movie.vote_count} {movie.vote_count !== 1 ? "votes" : "vote"}
-                                        </StyledVotes>
-                                    </StyledRate>
-                                </StyledDetails>
-                            </StyledMovieLink>
-                        )
-                    })}
+                    <Tile moviesList={moviesList} genreList={genreList} />
                 </StyledMovies>
                 <Outlet />
             </>
@@ -70,8 +52,8 @@ const Movies = () => {
             loadingMovies === "loading" || loadingGeners === "loading" ?
                 <Loading /> :
                 loadingMovies === "error" || loadingGeners === "error" ?
-                <Error /> :
-                null
+                    <Error /> :
+                    null
         )
     }
 };
