+++ src/core/rootSaga.js
@@ -5,12 +5,14 @@ import { movieCreditsSaga } from "../features/movieDetails/movieCreditsSaga";
 import { personListSaga } from "../features/personList/personListSaga";
 import { personDetailsSaga } from "../features/personDetails/personDetailsSaga";
 import { personCreditsSaga } from "../features/personDetails/personCreditsSaga";
+import { movieSearchParamsSaga } from "../features/searchParams/searchParamsSaga";
 
 export default function* rootSaga() {
   yield all([
     movieDetailsSaga(),
     movieListSaga(),
     movieCreditsSaga(),
+    movieSearchParamsSaga(),
     personListSaga(),
     personDetailsSaga(),
     personCreditsSaga(),
