+++ src/core/status/Loading/index.js
@@ -4,9 +4,9 @@ import { useSearchParams } from "react-router-dom";
 import { searchQueryParamName } from "../../../features/useQueryParameter";
 
 export const Loading = () => {
-    const [searchParams] = useSearchParams({[searchQueryParamName]: "" });
-    const query = searchParams.get(searchQueryParamName);
-    const content = query ? `Search results for "${query}"` : null;
+        const [searchParams] = useSearchParams({ [searchQueryParamName]: "" });
+        const query = searchParams.get(searchQueryParamName);
+        const content = query ? `Search results for "${query}"` : null;
 
     return (
         <>
