+++ src/core/store.js
@@ -4,6 +4,7 @@ import rootSaga from "./rootSaga";
 import movieDetailsReducer from "../features/movieDetails/movieDetailsSlice";
 import movieListReducer from "../features/moveList/movieListSlice";
 import genreListReducer from "../features/genreList/genreListSlice";
+import movieListSearchReducer from "../features/searchParams/searchParamsSlice";
 import movieCreditsReducer from "../features/movieDetails/movieCreditsSlice";
 import personListReducer from "../features/personList/personListSlice";
 import personDetailsReducer from "../features/personDetails/personDetailsSlice";
@@ -16,6 +17,7 @@ const store = configureStore({
     movieDetails: movieDetailsReducer,
     movieList: movieListReducer,
     genreList: genreListReducer,
+    movieListSearch: movieListSearchReducer,
     movieCredits: movieCreditsReducer,
     personList: personListReducer,
     personDetails: personDetailsReducer,
