+++ src/features/searchParams/searchParamsAPI.js
@@ -1,0 +1,10 @@
+import axios from "axios";
+
+const searchParamsAPI = (movieId) => {
+  return `https://api.themoviedb.org/3/search/movie?api_key=25dc6b875063fb2a5cdc0fd9237d002d&language=en-US&${movieId !== "" ? `query=${movieId}` : ""}&page=1&include_adult=false`;
+}
+
+
+export const getSearchParamsApi = (movieId) => {
+  return axios.get(searchParamsAPI(movieId)).then((response) => response.data);
+}
